requirements.txt
# ── Web Framework ──────────────────────────────────────────────────────────────
fastapi>=0.135.0
uvicorn[standard]>=0.42.0
python-multipart>=0.0.22        # FastAPI 文件上传必须

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.21.0              # SQLite 异步驱动

# ── Data Validation ────────────────────────────────────────────────────────────
pydantic>=2.11.0
pydantic-settings>=2.10.0

# ── Authentication ─────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.5.0   # JWT 签发与验证
bcrypt>=5.0.0                      # 密码哈希

# ── Rate Limiting ──────────────────────────────────────────────────────────────
slowapi>=0.1.9

# ── Caching ────────────────────────────────────────────────────────────────────
cachetools>=7.0.0              # 内存 TTLCache

# ── AI / LLM ───────────────────────────────────────────────────────────────────
openai>=2.0.0                  # AsyncOpenAI（DeepSeek/Qwen/Groq 均兼容 OpenAI 接口）
httpx>=0.28.0                  # LLM 代理请求 + TestClient
sentence-transformers>=5.0.0   # 本地文本嵌入（all-MiniLM-L6-v2）
numpy>=2.0.0                   # 向量相似度计算

# ── Agent Framework ────────────────────────────────────────────────────────────
crewai>=1.12.0
crewai-tools>=1.12.0

# ── Document Processing ────────────────────────────────────────────────────────
pdfplumber>=0.11.0             # PDF 文本提取（主）
PyMuPDF>=1.26.0                # PDF 文本提取（备）+ 图片渲染
Pillow>=12.0.0                 # 图片格式转换（简历图片）
rapidocr-onnxruntime>=1.4.0    # 图片/扫描件 OCR 兜底
python-docx>=1.2.0             # Word 报告导出
reportlab>=4.4.0               # PDF 报告导出

# ── Graph ──────────────────────────────────────────────────────────────────────
networkx>=3.6                  # 岗位知识图谱

# ── Template Engine ────────────────────────────────────────────────────────────
jinja2>=3.1.0                  # LLM 提示词 Jinja2 模板

# ── Scheduler ──────────────────────────────────────────────────────────────────
apscheduler>=3.11.0            # 定时任务（JD抓取/趋势刷新）
